requirements.txt
docopt>=0.6,<0.7
fusepy==2.0.4